requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==2.0
fonttools==4.31.1
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mpmath==1.2.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sympy==1.10.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
